% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{chartOutput}
\alias{chartOutput}
\title{Chart output for shiny in ui.R}
\usage{
chartOutput(outputId)
}
\arguments{
\item{outputId}{outputId rendered from server.R}
}
\description{
Chart output for shiny in ui.R
}

