% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{create_dyadid}
\alias{create_dyadid}
\title{Create a dyad identifier variable}
\usage{
create_dyadid(data, name, dyad_mem1, dyad_mem2, directed = F)
}
\arguments{
\item{data}{A \code{data.frame} object containing individual dyad member ID variables.}

\item{name}{A string for the name of new dyad ID variable to be created.}

\item{dyad_mem1}{A string for the name of first dyad member ID variable.}

\item{dyad_mem2}{A string for the name of second dyad member ID variable.}

\item{directed}{A logical value indicating whether the dyad ID should be for the directed or undirected dyad.}
}
\value{
The input \code{data.frame} with a new dyad ID variable appended to it.
}
\description{
This function returns a \code{data.frame} with a new dyad ID variable based
on a provided \code{data.frame} with individual dyad member ID variables.
}
