% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{dc_code}
\alias{dc_code}
\title{Wrap a character string of javascript code or function definition}
\usage{
dc_code(x)
}
\arguments{
\item{x}{the character string}
}
\description{
Wrap a character string of javascript code or function definition
}

