% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dc_filters}
\alias{dc_filters}
\title{Display current filters}
\usage{
dc_filters(text = "Current filter:")
}
\arguments{
\item{text}{text put before the displayed filters}
}
\description{
Display current filters
}
\examples{
\dontrun{

# For example, we put following in the ui.R
div(dc_filters(), dc_reset("mychart"), id = "mychart")
}
}

