% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr.R
\name{dc_reset}
\alias{dc_reset}
\title{Add link to reset filter}
\usage{
dc_reset(id, name = "reset")
}
\arguments{
\item{id}{chart id the reset link used to reset}

\item{name}{what the link appears as, default is "reset", used in ui.R}
}
\description{
Add link to reset filter
}

