% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dc_resetAll}
\alias{dc_resetAll}
\title{Add link to reset all filters}
\usage{
dc_resetAll(name = "Reset All")
}
\arguments{
\item{name}{what the link appears as, default is "reset", used in ui.R}
}
\description{
Add link to reset all filters
}

