% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dc_stack}
\alias{dc_stack}
\title{Add a stacked chart}
\usage{
dc_stack(reduce, name = NULL, accessor = NULL)
}
\arguments{
\item{reduce}{reduce function}

\item{name}{name appearing in the label}

\item{accessor}{accssor function}
}
\description{
Add a stacked chart
}

