% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dc_ui}
\alias{dc_ui}
\title{Create a division in ui.R showing one dc chart}
\usage{
dc_ui(id, text = "", reset = TRUE, show_filter = FALSE,
  reset_text = "reset", filter_text = "Current Filter:")
}
\arguments{
\item{id}{the chart id}

\item{reset}{whether or not show the reset link}

\item{show_filter}{whether or not show the current filters}

\item{reset_text}{the text for reset link}

\item{filter_text}{the text preceding displayed filters}

\item{name}{the text for the chart showing above the chart}
}
\description{
This function combines frequently used UI such as current filters,
reset link, and chart names
}
\examples{
\dontrun{

# We can put following in the ui.R
dc_ui("chart1", "The first chart", show_filter = TRUE, filter_text = "Selected:")
}
}

