% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{dcr}
\alias{dcr}
\title{Create dc visulization object}
\usage{
dcr(data, charts = list())
}
\arguments{
\item{data}{the data used to visulize}

\item{charts}{a list of dc charts in the visulization}
}
\description{
Create dc visulization object
}

