% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_funcs.R
\name{dcr_sandwich}
\alias{dcr_sandwich}
\title{Compute dyadic clustering robust (DCR) variance estimates (\code{sandwich} version)}
\usage{
dcr_sandwich(
  model,
  dyad_id,
  dyad_mem1,
  dyad_mem2,
  data,
  posdef = TRUE,
  dofcorr = TRUE
)
}
\arguments{
\item{model}{A model object.}

\item{dyad_mem1}{A string for the name of first dyad member identifier variable.}

\item{dyad_mem2}{A string for the name of second dyad member identifier variable.}

\item{data}{A \code{data.frame} object containing dyadic data.}

\item{posdef}{A logical value indicating whether or not DCR variance-covariance matrix is forced to be positive semi-definite.}

\item{dofcorr}{A logical value indicating whether or not to apply a small sample correction to the final DCRSE estimates. This correction is equivalent to multiplying DCRSEs by sqrt(N/(N-1)), where N is the number of unique dyad members in the analytic sample. Correspondingly, when computing p-values, the test statistic should be compared to a t-distribution with DOF = N - 1.}
}
\value{
A list containing DCRSEs and DCR variances for model parameters, as well as the number of unique dyad members in the analytic sample.
}
\description{
This function computes DCR variance estimates and standard errors
(DCRSEs) for model parameters.
}
