% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{dcrchart}
\alias{dcrchart}
\title{Function to create a dc chart}
\usage{
dcrchart(type, id, dimension, reduce, width, height, ..., group_name = "")
}
\arguments{
\item{type}{chart type such as pieChart, barChart etc}

\item{id}{chart div id}

\item{dimension}{chart dimension}

\item{reduce}{chart reduce function}

\item{width}{chart width, e.g. 400}

\item{height}{chart height, e.g. 600}

\item{...}{additional chart options}

\item{group_name}{group name for this chart, used for name appeared in legend}
}
\description{
Function to create a dc chart
}

