% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr.R
\name{geojson}
\alias{geojson}
\title{Define overlayGeoJson for geoChoropleth chart}
\usage{
geojson(jsonfile, name, keyAccessor)
}
\arguments{
\item{jsonfile}{path for geojson file}

\item{name}{name of the layer}

\item{keyAccessor}{accessor function used to extract "key" from GeoJson data}
}
\description{
Define overlayGeoJson for geoChoropleth chart
}

