% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{html}
\alias{html}
\title{Convert dcr object to html code}
\usage{
html(object, divs = FALSE, csv = FALSE, filename = NULL,
  input_binding = FALSE)
}
\arguments{
\item{divs}{whether or not to generate divisions}

\item{csv}{whether or not export data in csv format, otherwise it will export as json format}

\item{obj}{dcr chart object}

\item{csv}{file filename}

\item{input_biding}{whether or not binding filter values to a input so server.R can
access the filters of each chart through input$chart}
}
\description{
Convert dcr object to html code
}

