% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{js}
\alias{js}
\title{Convert R object to javascript object}
\usage{
js(x)
}
\arguments{
\item{x}{R object, if x is a string wrapped by \code{\link{dc_code}} then it is converted to code
such as javascript function definition}
}
\description{
Convert R object to javascript object
}

