% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{parse_chart_height}
\alias{parse_chart_height}
\title{Determine if chart height is given by num of px or % of window height}
\usage{
parse_chart_height(height)
}
\arguments{
\item{height}{The user specified height}
}
\description{
Determine if chart height is given by num of px or % of window height
}

