% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{parse_chart_width}
\alias{parse_chart_width}
\title{Determine if chart width is given by num of px or % of window width}
\usage{
parse_chart_width(width)
}
\arguments{
\item{width}{The user specified width}
}
\description{
Determine if chart width is given by num of px or % of window width
}

