% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduceCountPercentOfTotal}
\alias{reduceCountPercentOfTotal}
\title{Reduce function to calculate the count of variable based
divided by the given total. Result is a percent.}
\usage{
reduceCountPercentOfTotal(var, total)
}
\arguments{
\item{var}{the name of the variable to calculate percent.}

\item{total}{The total number used when calculating the percent.}
}
\description{
Reduce function to calculate the count of variable based
divided by the given total. Result is a percent.
}

