% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduceMean}
\alias{reduceMean}
\title{Reduce function to calculate mean of a variable for each group}
\usage{
reduceMean(var, weight = NULL)
}
\arguments{
\item{var}{the name of the variable to calculate mean}

\item{weight}{variable specifying weight when calculating mean}
}
\description{
Reduce function to calculate mean of a variable for each group
}

