% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduceSum}
\alias{reduceSum}
\title{Reduce function to calculate sum of a variable for each group}
\usage{
reduceSum(var)
}
\arguments{
\item{var}{the name of the variable to sum over}
}
\description{
Reduce function to calculate sum of a variable for each group
}

