% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{responsive_width_charts_js_var}
\alias{responsive_width_charts_js_var}
\title{Adds the charts with responsive widths to JS object
so that responsive_chart_manager.js will update them.}
\usage{
responsive_width_charts_js_var(charts)
}
\arguments{
\item{charts}{List of all charts.}
}
\description{
Adds the charts with responsive widths to JS object
so that responsive_chart_manager.js will update them.
}

