% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_fun}
\alias{simple_fun}
\title{Simple javascript function wrapper}
\usage{
simple_fun(x)
}
\arguments{
\item{x}{expression to return}
}
\description{
generate simple one argument javascript function definition on d as
function(d) {return x;}, where x is some expression on d for function return
}

