% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{use_dimension}
\alias{use_dimension}
\title{Use other chart's dimension as the chart's dimension}
\usage{
use_dimension(id)
}
\arguments{
\item{the}{id of the chart whose dimension will be used}
}
\description{
Use other chart's dimension as the chart's dimension
}

