% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{x_linear}
\alias{x_linear}
\title{Generate linear scale used with x option}
\usage{
x_linear(xlim)
}
\arguments{
\item{xlim}{x axis range, should be a numeric vector of length 2}
}
\description{
Generate linear scale used with x option
}

