% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{x_ordinal}
\alias{x_ordinal}
\title{Generate ordinal scale used with x option}
\usage{
x_ordinal(levels = NULL)
}
\arguments{
\item{levels}{order of levels, should be a vector of specified ordered levels}
}
\description{
Generate ordinal scale used with x option
}

