% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_funcs.R
\name{dcr}
\alias{dcr}
\title{Compute dyadic clustering robust (DCR) variance estimates}
\usage{
dcr(model, dyad_id, dyad_mem1, dyad_mem2, data, posdef = TRUE)
}
\arguments{
\item{model}{Model object (note: must be compatible with \code{sandwich} package functions)}

\item{dyad_id}{String for name of dyad identifier variable}

\item{dyad_mem1}{String for name of first dyad member identifier variable}

\item{dyad_mem2}{String for name of second dyad member identifier variable}

\item{data}{\code{data.frame} object containing dyadic data, including dyad and dyad member identifier variables}

\item{posdef}{Logical value indicating whether or not DCR variance-covariance matrix is forced to be positive semi-definite}
}
\value{
A list containing DCR standard errors and variances for model parameters
}
\description{
This function computes DCR standard error and variance estimates for
model parameters.
}
