% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_funcs.R
\name{dcr}
\alias{dcr}
\title{Compute dyadic-clustering robust (DCR) variance estimates}
\usage{
dcr(model, dyad_id, dyad_mem1, dyad_mem2, data)
}
\arguments{
\item{model}{Model object (must be compatible with Sandwich package functions)}

\item{dyad_id}{String for variable name for dyad identifier}

\item{dyad_mem1}{String for variable name for first dyad member}

\item{dyad_mem2}{String for variable name for second dyad member}

\item{data}{Data frame object containing dyadic data with dyad identifier variables}
}
\value{
A list containing DCR standard errors variances for model parameters
}
\description{
This function computes DCR standard error and variance estimates for
model parameters.
}
