% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_funcs.R
\name{dcr_custom}
\alias{dcr_custom}
\title{Compute dyadic-clustering robust (DCR) variance estimates (customizable)}
\usage{
dcr_custom(model, dyad_id, dyad_mem1, dyad_mem2, spec_vars, data, ef = NULL)
}
\arguments{
\item{model}{Model object}

\item{dyad_id}{String for variable name for dyad identifier}

\item{dyad_mem1}{String for variable name for first dyad member}

\item{dyad_mem2}{String for variable name for second dyad member}

\item{spec_vars}{String vector of variable names in model specification}

\item{data}{Data frame object containing dyadic data with dyad identifier variables}

\item{ef}{Optional matrix containing empirical estimating functions}
}
\value{
A list containing DCR standard errors and variances for model parameters
}
\description{
This function computes DCR standard error and variance estimates for
model parameters with more model object handling flexibility than dyad_robust.
}
