% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_funcs.R
\name{dcr_parallel}
\alias{dcr_parallel}
\title{Compute dyadic-clustering robust (DCR) variance estimates (parallelized)}
\usage{
dcr_parallel(
  model,
  dyad_id,
  dyad_mem1,
  dyad_mem2,
  ncore = ceiling(detectCores()/2),
  data
)
}
\arguments{
\item{model}{Model object (must be compatible with Sandwich package functions)}

\item{dyad_id}{String for variable name for dyad identifier}

\item{dyad_mem1}{String for variable name for first dyad member}

\item{dyad_mem2}{String for variable name for second dyad member}

\item{ncore}{Integer specifying the number of cores to be used for parallel computation}

\item{data}{Data frame object containing dyadic data with dyad identifier variables}
}
\value{
A list containing DCR standard errors variances for model parameters
}
\description{
This function enables the parallel computation of DCR standard error
and variance estimates.
}
\details{
The default number of cores is set to the ceiling of half the total number of cores
available.
}
