% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcr_classfun.R
\name{html}
\alias{html}
\title{Convert dcr object to html code}
\usage{
html(object, divs = FALSE)
}
\arguments{
\item{divs}{whether or not to generate divisions}

\item{obj}{dcr chart object}
}
\description{
Convert dcr object to html code
}

