% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderChart}
\alias{renderChart}
\title{Render dc chart in server.R}
\usage{
renderChart(expr, divs = FALSE, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression to evaluate, it should return a dcr chart object}

\item{divs}{whether to generate division automatically in ui}

\item{env}{environment to evaluate}

\item{quoted}{quoted expression or not}
}
\description{
Render dc chart in server.R
}

