% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderChart_csv}
\alias{renderChart_csv}
\title{Render dc chart in server.R}
\usage{
renderChart_csv(session, expr, divs = FALSE, env = parent.frame(),
  quoted = FALSE)
}
\arguments{
\item{session}{shiny session}

\item{expr}{expression to evaluate, it should return a dcr chart object}

\item{divs}{whether to generate division automatically in ui}

\item{env}{environment to evaluate}

\item{quoted}{quoted expression or not}
}
\description{
Render dc chart in server.R
}

