% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x_time}
\alias{x_time}
\title{Generate time scale used with x option}
\usage{
x_time(daterange)
}
\arguments{
\item{daterange}{date range to display, should be a Date vector of length 1}
}
\description{
Generate time scale used with x option
}

